requirements.txt
python-etcd
jinja2
bpython
